requirements.txt
flask
flask-wtf
Flask-SQLAlchemy
Flask-Migrate
sqlalchemy-utils
python-dotenv
psycopg2-binary